Update/RecyclerView

+++ 02-Firebase/app/src/main/java/com/example/proyectofirebase/CProducto.kt
@@ -10,26 +10,29 @@ import com.google.firebase.ktx.Firebase
 import io.grpc.internal.DnsNameResolver
 
 class CProducto : AppCompatActivity() {
+
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_cproducto)
 
         val botonCrear = findViewById<Button>(R.id.btn_crear_producto)
-        botonCrear.setOnClickListener {
-            crearProducto()
-        }
-
+        botonCrear
+            .setOnClickListener {
+                crearProducto()
+            }
     }
 
-    fun crearProducto(){
+    fun crearProducto() {
         val editTextNombre = findViewById<EditText>(R.id.et_nombre_producto)
-        val editTextPrecio  = findViewById<EditText>(R.id.et_precio_producto)
+        val editTextPrecio = findViewById<EditText>(R.id.et_precio_producto)
         val nuevoProducto = hashMapOf<String, Any>(
             "nombre" to editTextNombre.text.toString(),
-            "precio" to editTextPrecio.text.toString()
+            "precio" to editTextPrecio.text.toString().toDouble()
         )
         val db = Firebase.firestore
         val referencia = db.collection("producto")
+
 
         referencia
             .add(nuevoProducto)
@@ -37,7 +40,7 @@ class CProducto : AppCompatActivity() {
                 editTextNombre.text.clear()
                 editTextPrecio.text.clear()
             }
-            .addOnFailureListener{}
+            .addOnFailureListener { }
 
     }
